+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 ### Changed
 - The Tweet field now returns a tweet URL instead of a Tweet model.
 - Renamed `getTweetById()` and `getTweetByUrl()` methods to `getTweet()`.
+- Removed unused `dukt\twitter\helpers\TwitterHelper::formatTime()` method.
 
 ### Fixed
 - Fixed a bug where the Tweet field would throw an error when trying to show a deleted tweet. ([#7](https://github.com/dukt/twitter/issues/7))